requirements.txt
playwright>=1.40.0
apscheduler>=3.10.4
loguru>=0.7.2
PyYAML>=6.0.1
pytest>=7.4.0
pytest-asyncio>=0.23.0
